aider: Set `shfmt` as a fixer for Bash scripts and configure its options globally.
# Aider chat conversation:

USER: :ALEFix is not reformatting my bash script. What am I doing wrong in the bash.vim file?

ASSISTANT: It seems that the `bash.vim` file is not setting up ALE to use any fixers for Bash scripts. To enable `shfmt` as a fixer for Bash scripts, you need to add it to the `b:ale_fixers` variable for the `sh` filetype. Here's the diff to update the `bash.vim` file:

```diff
--- .vim/pack/settings/start/settings/after/ftplugin/bash.vim
+++ .vim/pack/settings/start/settings/after/ftplugin/bash.vim
@@ ... @@
 setlocal shiftwidth=2
 setlocal softtabstop=2
 setlocal expandtab
 setlocal autoindent
 setlocal smartindent

-" coc.nvim is handling the LSP (including linters) so we don't need to setup
-" shellcheck here.
+let b:ale_fixers = ['shfmt']

-let b:ale_fixers = ['shfmt']
+let g:ale_sh_shfmt_options =
+      \ '--indent 2 ' .
+      \ '--binary-next-line ' .
+      \ '--case-indent ' .
+      \ '--space-redirects '
```

This change ensures that `shfmt` is set as a fixer for Bash scripts and that the options for `shfmt` are configured globally.