aider: Update ALE to run shfmt in the directory of the current file and then return to the original directory.
# Aider chat conversation:

USER: Are any of these files setting an option to not change the working directory to the location of the file being edited?

ASSISTANT: Yes, the `.vim/pack/settings/start/settings/plugin/set.vim` file contains a setting that prevents Vim from changing the working directory to the location of the file being edited. The option responsible for this behavior is `autochdir`. When this option is not set or set to `noautochdir`, Vim does not change the current working directory to the directory of the file being edited.

Here is the relevant line from the `set.vim` file:

```vim
set noautochdir
```

This line explicitly sets the `noautochdir` option, ensuring that the working directory remains unchanged when opening files in different directories. If you want to change this behavior, you would need to set the `autochdir` option instead. However, since you've mentioned that the files are read-only, no changes can be proposed or made to this file unless you decide to make it read-write.

USER: Can we configure ale to change to the directory the current file is in, run shfmt, then change back to the original directory so that :ALEFix will work?

ASSISTANT: To configure ALE to change to the directory of the current file, run `shfmt`, and then change back to the original directory, you can use ALE's `g:ale_fixers` configuration. You can define a custom fixer function that changes the directory before running `shfmt` and then changes back afterwards.

Here's the diff to update the `.vim/pack/settings/start/settings/after/ftplugin/sh.vim` file to include a custom fixer function for `shfmt`:

```diff
--- .vim/pack/settings/start/settings/after/ftplugin/sh.vim
+++ .vim/pack/settings/start/settings/after/ftplugin/sh.vim
@@ ... @@
 " coc.nvim is handling the LSP (including linters) so we don't need to setup
 " shellcheck here.

-let b:ale_fixers = ['shfmt']
+function! s:shfmt_fixer(buffer) abort
+    let l:cwd = getcwd()
+    let l:filename = expand('#' . a:buffer . ':p')
+    let l:dirname = fnamemodify(l:filename, ':h')
+    execute 'cd' l:dirname
+    let l:fixer_result = ale#fixers#shfmt#Fix(a:buffer)
+    execute 'cd' l:cwd
+    return l:fixer_result
+endfunction
+
+let b:ale_fixers = {'sh': function('s:shfmt_fixer')}

 let g:ale_sh_shfmt_options =
       \ '--write ' .
```

This custom fixer function `s:shfmt_fixer` will change the directory to the one containing the file being edited, run `shfmt` using ALE's built-in `shfmt` fixer, and then change back to the original directory. The `b:ale_fixers` variable is updated to use this custom function for files with the `sh` filetype.